waterMarkForPDF/requirements.txt
python-pptx
Pillow
aspose-slides